+++ src/common/Environment.ts
@@ -62,6 +62,10 @@ args
     ["", "skipApiVersionCheck"],
     "Optional. Skip the request API version check, request with all Api versions will be allowed"
   )
+  .option(
+    ["", "disableProductStyleUrl"],
+    "Optional. Disable getting account name from the host of request Uri, always get account name from the first path segment of request Uri."
+  )
   .option(["", "oauth"], 'Optional. OAuth level. Candidate values: "basic"')
   .option(["", "cert"], "Optional. Path to certificate file")
   .option(["", "key"], "Optional. Path to certificate key .pem file")
@@ -124,6 +128,14 @@ export default class Environment implements IEnvironment {
       return true;
     }
     // default is false which will check API veresion
+    return false;
+  }
+
+  public disableProductStyleUrl(): boolean {
+    if (this.flags.disableProductStyleUrl !== undefined) {
+      return true;
+    }
+    // default is false which will try to get account name from request Uri hostname
     return false;
   }