+++ src/common/VSCServerManagerBlob.ts
@@ -86,7 +86,8 @@ export default class VSCServerManagerBlob extends VSCServerManagerBase {
       env.cert(),
       env.key(),
       env.pwd(),
-      env.oauth()
+      env.oauth(),
+      env.disableProductStyleUrl()
     );
     return config;
   }