+++ src/common/VSCServerManagerQueue.ts
@@ -97,7 +97,8 @@ export default class VSCServerManagerBlob extends VSCServerManagerBase {
       env.cert(),
       env.key(),
       env.pwd(),
-      env.oauth()
+      env.oauth(),
+      env.disableProductStyleUrl()
     );
     return config;
   }