+++ src/queue/main.ts
@@ -64,7 +64,8 @@ async function main() {
     env.cert(),
     env.key(),
     env.pwd(),
-    env.oauth()
+    env.oauth(),
+    env.disableProductStyleUrl()
   );
 
   // We use logger singleton as global debugger logger to track detailed outputs cross layers