+++ src/table/TableConfiguration.ts
@@ -34,7 +34,8 @@ export default class TableConfiguration extends ConfigurationBase {
     cert: string = "",
     key: string = "",
     pwd: string = "",
-    oauth?: string
+    oauth?: string,
+    disableProductStyleUrl: boolean = false
   ) {
     super(
       host,
@@ -48,7 +49,8 @@ export default class TableConfiguration extends ConfigurationBase {
       cert,
       key,
       pwd,
-      oauth
+      oauth,
+      disableProductStyleUrl
     );
   }
 }