+++ ChangeLog.md
@@ -7,6 +7,7 @@
 General:
 
 - Restore ability to connect to host.docker.internal.
+- Added new parameter `--disableProductStyleUrl`, to force parse storage account from request Uri path, instead of from request Uri host.
 
 Blob: